refactor(stark-ui): move internal properties from StarkProgressIndicatorConfig to a new StarkProgressIndicatorFullConfig
BREAKING CHANGES: `StarkProgressIndicatorConfig` object now contains
only `topic` and `type` properties. The rest of properties (`visible`, `listenersCount` and `pendingListenersCount`) are now in a new `StarkProgressIndicatorFullConfig` object that is only used internally by the `StarkProgressIndicatorService`. This change does not affect the functionality of the `StarkProgressIndicatorDirective` since the removed properties were in fact not used by the directive.